% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataClass.R
\docType{class}
\name{cohortData}
\alias{cohortData}
\title{Data Storage Class for cohort dataset}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{cohortData} object
}
\description{
Class that defines the standard format for the cohort dataset, i.e. the table
that specifies for each subject in the cohort:
\enumerate{
\item a unique subject identifier,
\item the date of study entry,
\item the date of end of follow-up,
\item the reason for end of follow-up (failure or right-censoring), and
\item baseline measurements of time-dependent or time-independent covariates.
}
}
\section{Fields}{

\describe{
\item{\code{data}:}{\code{data.table} containing
the input cohort dataset to be wrapped in and processed.
The table must contain a single row for each subject in the cohort.
Cannot have columns named 'IDvar', 'index_date', 'EOF_date',
'EOF_type', or 'L0'.
}
\item{\code{IDvar}:}{\code{character} providing the name of the column of
\code{data} that contains the unique subject identifier.
}
\item{\code{index_date}:}{\code{character} providing the name of the
column of \code{data} that contains the date of study entry.
}
\item{\code{EOF_date}:}{\code{character} providing the name of the column
of \code{data} that contains the date of end of follow-up. All observations
with the end of follow-up date equal to the study entry date will be ignored
(i.e., excluded from the cohort).
}
\item{\code{EOF_type}:}{\code{character} providing the name of the column
of \code{data} corresponding to the reason for end of follow-up.
}
\item{\code{Y_name}:}{\code{character} or \code{integer}  providing the
unique value in column \code{EOF_type} that encodes the end of
follow-up due to failure (i.e., occurrence of the outcome event of
interest).
}
\item{\code{L0}:}{vector of \code{character} providing the names of the
columns of \code{data} that contain baseline covariate
measurements. Covariate values must be encoded by a character or
numeric vector (e.g., factors are not allowed).
}
\item{\code{L0_timeIndep}:}{named list specifying, for each time-independent
covariates in \code{L0}, a sublist with only the following three named
elements:
\enumerate{
\item \code{categorical}: \code{logical} indicating
whether the time-independent covariate is
continuous ('FALSE') or categorical ('TRUE'). Cannot be missing.
\item \code{impute}: \code{character} specifying the imputation
method for
missing measurements of the time-independent covariate. Possible values are
'default', 'mean', 'mode', 'median'.
If missing, imputation with
the 'mean' and 'mode' is used for continuous and categorical covariates,
respectively. Imputation with 'mean', 'mode', or 'median' is based on
measurements in \code{data} from subjects with observed covariate values.
'mean' and 'median' can only be used for continuous covariates.
'mode' can only be used for categorical covariates. Imputation with 'default'
replaces missing values with 0 if the covariate is numeric and with 'Unknown'
otherwise.
\item \code{impute_default_level}: \code{character} or
\code{numeric} specifying the
imputation value to be used when \code{impute}='default'. The value must be
a length 1 \code{character} (resp. \code{numeric}) for a covariate encoded by a
\code{character} (resp. \code{numeric}) vector.
If missing, the default
values 0 and 'Unknown' are used for numeric and character covariates,
respectively.
}
Each element of the list \code{L0_timeIndep} must be named with the time-independent
covariate in \code{L0} to which the sublist information applies. \code{L0_timeIndep}
can be missing if there is no time-independent covariate in \code{data}.
}
}
}

\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{cohortData$new()}}
\item \href{#method-clone}{\code{cohortData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cohortData$new(
  data,
  IDvar,
  index_date,
  EOF_date,
  EOF_type,
  Y_name,
  L0,
  L0_timeIndep
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cohortData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
