% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataClass.R
\docType{class}
\name{expData}
\alias{expData}
\title{Data Storage Class for exposure dataset}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{expData} object
}
\description{
Class that defines the standard format for the exposure dataset, i.e. the
table that specifies all follow-up time intervals during which a study
subject is exposed to an exposure level other than a reference level
specified by the analyst. For each episode of exposure to a non-reference
level, the table specifies:
\enumerate{
\item a unique subject identifier,
\item the exposure episode start date,
\item the exposure episode end date, and
\item the non-reference exposure level (required only when the exposure is not
binary).
}
}
\section{Fields}{

\describe{
\item{\code{data}:}{\code{data.table} containing
the input exposure dataset to be wrapped in and processed.
The table can contain multiple rows per subject. There
should be no row for subjects who are only
exposed to the reference exposure level during follow-up.
Exposure levels must be encoded by a character vector or an integer
vector. There cannot be any overlapping exposure episodes. Cannot
contain missing values. Cannot have columns named 'IDvar',
'start_date', 'end_date', or 'exp_level'.
}
\item{\code{IDvar}:}{\code{character} providing the name of the column of
\code{data} that contains the unique subject identifier.
}
\item{\code{start_date}:}{\code{character} providing the name of the
column of \code{data} that contains the exposure episode start
date.
}
\item{\code{end_date}:}{\code{character} providing the name of the column
of \code{data} that contains the exposure episode end date.
}
\item{\code{exp_level}:}{\code{character} providing the name of the
column of \code{data} that contains the non-reference exposure
level. Can be missing only if \code{exp_ref} is also missing. If
missing, the exposure is assumed to be binary and its reference
level is encoded by 0.
}
\item{\code{exp_ref}:}{\code{character} or \code{integer} identifying the
exposure reference level. Cannot be a value of the exp_level column
of data. Can be missing only if \code{exp_level} is also missing.
}
}
}

\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{expData$new()}}
\item \href{#method-checkAgainst}{\code{expData$checkAgainst()}}
\item \href{#method-clone}{\code{expData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{expData$new(data, IDvar, start_date, end_date, exp_level = NA, exp_ref = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkAgainst"></a>}}
\if{latex}{\out{\hypertarget{method-checkAgainst}{}}}
\subsection{Method \code{checkAgainst()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{expData$checkAgainst(otherData)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{expData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
