% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataClass.R
\docType{class}
\name{instExpData}
\alias{instExpData}
\title{Data Storage Class for instant exposure dataset}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{instExpData} object
}
\description{
Class that defines the standard format for the instant exposure dataset, i.e. the
table that specifies all follow-up times during which a study
subject is exposed to a non-reference exposure level. For each instantaneous
exposure, the table specifies:
\enumerate{
\item a unique subject identifier,
\item the exposure time,
\item the exposure level (required only when the exposure is not
binary).
}
}
\section{Fields}{

\describe{
\item{\code{data}:}{\code{data.table} containing
the input exposure dataset to be wrapped in and processed.
The table can contain multiple rows per subject. There
should be no row for subjects who are only
exposed to the reference exposure level during follow-up.
Exposure levels must be encoded by one or more character, integer,
or numeric vector(s). Multiples rows per subject and time are permitted.
Cannot
contain missing values. Cannot have columns named 'IDvar',
'start_date', or 'exp_level'.
}
\item{\code{IDvar}:}{\code{character} providing the name of the column of
\code{data} that contains the unique subject identifier.
}
\item{\code{start_date}:}{\code{character} providing the name of the
column of \code{data} that contains the exposure
date.
}
\item{\code{exp_level}:}{\code{character} vector providing the name(s) of the
column(s) of \code{data} that contain(s) the non-reference exposure
level(s). Can be missing if there is only one non-reference exposure level.
If missing, the exposure is assumed to be binary and its reference
level is encoded by 0.
}
}
}

\keyword{data}
\section{Super class}{
\code{\link[LtAtStructuR:expData]{LtAtStructuR::expData}} -> \code{instExpData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{instExpData$new()}}
\item \href{#method-checkAgainst}{\code{instExpData$checkAgainst()}}
\item \href{#method-clone}{\code{instExpData$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{instExpData$new(data, IDvar, start_date, exp_level = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkAgainst"></a>}}
\if{latex}{\out{\hypertarget{method-checkAgainst}{}}}
\subsection{Method \code{checkAgainst()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{instExpData$checkAgainst(otherData)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{instExpData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
