% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataClass.R
\docType{class}
\name{LtAtData}
\alias{LtAtData}
\title{Data Storage Class for an LtAt dataset}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{LtAtData} object
}
\description{
Class that defines the specifications for constructing an LtAt dataset and the dataset itself.
}
\section{Fields}{

\describe{
\item{\code{data}:}{\code{data.table} containing
the LtAt dataset. NULL by default before construction.
}
\item{\code{cohort_data}:}{\code{cohortData} specifying the cohort dataset.
}
\item{\code{exp_data}:}{\code{expData} specifying the time-varying exposure dataset.
Its \code{IDvar} field must contain the same value as that of the
\code{cohort_data} (i.e.,
the same column name must be used for the ID column of the exposure and cohort
datasets). The unique subject identifiers in the \code{data} field of \code{exp_data}
must be a subset of those founds in the \code{data} field of \code{cohort_data} (i.e.,
all IDs in the exposure dataset must also be present in the cohort dataset.)
An exposure measurement stored in the \code{data} field of \code{exp_data}
will be ignored
if its date is either strictly before the subject's index date stored in the
\code{data} field of \code{cohort_data} or strictly after the subject's end of
follow-up date stored in the \code{data} field of \code{cohort_data} (i.e., only
exposure measurements collected between the start and end of follow-up will be
retained).
}
\item{\code{cov_data}:}{a list of \code{timeDepCovData} that
specifies time-varying covariate(s). Each element of the list must have a distinct
value stored in its \code{L_name} field and this value must also be found in the
\code{L0} field of \code{cohort_data} (i.e., time-varying covariates cannot have
the same name and must have their baseline measurements stored in the cohort dataset).
The \code{IDvar} field of each element \code{timeDepCovData} of \code{cov_data}
must contain the same value as
that of the \code{cohort_data} (i.e.,
the same column name must be used for the ID column of each covariate dataset and
that of the cohort dataset). The unique subject identifiers in the \code{data}
field of each element
\code{timeDepCovData} of \code{cov_data}
must be a subset of those founds in the \code{data} field of \code{cohort_data} (i.e.,
all IDs in a covariate dataset must also be present in the cohort dataset.)
A covariate measurement stored in the \code{data} field of an element
\code{timeDepCovData}
of \code{cov_data} will be ignored
if its date is either strictly before the subject's index date stored in the
\code{data} field of \code{cohort_data} or strictly after the subject's end of
follow-up date stored in the \code{data} field of \code{cohort_data}. (i.e., only
covariate measurements collected between the start and end of follow-up will be
retained).
}
}
}

\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LtAtData$new()}}
\item \href{#method-addSpec}{\code{LtAtData$addSpec()}}
\item \href{#method-createIntervals}{\code{LtAtData$createIntervals()}}
\item \href{#method-assignAC}{\code{LtAtData$assignAC()}}
\item \href{#method-assignL}{\code{LtAtData$assignL()}}
\item \href{#method-imputeL}{\code{LtAtData$imputeL()}}
\item \href{#method-cleanUp}{\code{LtAtData$cleanUp()}}
\item \href{#method-clone}{\code{LtAtData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LtAtData$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addSpec"></a>}}
\if{latex}{\out{\hypertarget{method-addSpec}{}}}
\subsection{Method \code{addSpec()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LtAtData$addSpec(spec)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createIntervals"></a>}}
\if{latex}{\out{\hypertarget{method-createIntervals}{}}}
\subsection{Method \code{createIntervals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LtAtData$createIntervals(time_unit)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assignAC"></a>}}
\if{latex}{\out{\hypertarget{method-assignAC}{}}}
\subsection{Method \code{assignAC()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LtAtData$assignAC(firs_exp_rule = 1, exp_threshold = 0.5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assignL"></a>}}
\if{latex}{\out{\hypertarget{method-assignL}{}}}
\subsection{Method \code{assignL()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LtAtData$assignL(first_exp_rule = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-imputeL"></a>}}
\if{latex}{\out{\hypertarget{method-imputeL}{}}}
\subsection{Method \code{imputeL()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LtAtData$imputeL()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleanUp"></a>}}
\if{latex}{\out{\hypertarget{method-cleanUp}{}}}
\subsection{Method \code{cleanUp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LtAtData$cleanUp(format, dates)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LtAtData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
