% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataClass.R
\docType{class}
\name{LtAtDataInst}
\alias{LtAtDataInst}
\title{Data Storage Class for LtAt dataset based on instant exposure}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{LtAtDataInst} object
}
\description{
Class that defines the specifications for constructing an LtAt dataset and the dataset itself
using an instant exposure dataset.
}
\section{Fields}{

\describe{
\item{\code{data}:}{\code{data.table} containing
the LtAt dataset. NULL by default before construction.
}
\item{\code{cohort_data}:}{\code{cohortData} specifying the cohort dataset.
}
\item{\code{exp_data}:}{\code{expData} specifying the time-varying exposure dataset.
Its \code{IDvar} field must contain the same value as that of the
\code{cohort_data} (i.e.,
the same column name must be used for the ID column of the exposure and cohort
datasets). The unique subject identifiers in the \code{data} field of \code{exp_data}
must be a subset of those founds in the \code{data} field of \code{cohort_data} (i.e.,
all IDs in the exposure dataset must also be present in the cohort dataset.)
An exposure measurement stored in the \code{data} field of \code{exp_data}
will be ignored
if its date is either strictly before the subject's index date stored in the
\code{data} field of \code{cohort_data} or strictly after the subject's end of
follow-up date stored in the \code{data} field of \code{cohort_data} (i.e., only
exposure measurements collected between the start and end of follow-up will be
retained).
}
\item{\code{cov_data}:}{a list of \code{timeDepCovData} that
specifies time-varying covariate(s). Each element of the list must have a distinct
value stored in its \code{L_name} field and this value must also be found in the
\code{L0} field of \code{cohort_data} (i.e., time-varying covariates cannot have
the same name and must have their baseline measurements stored in the cohort dataset).
The \code{IDvar} field of each element \code{timeDepCovData} of \code{cov_data}
must contain the same value as
that of the \code{cohort_data} (i.e.,
the same column name must be used for the ID column of each covariate dataset and
that of the cohort dataset). The unique subject identifiers in the \code{data}
field of each element
\code{timeDepCovData} of \code{cov_data}
must be a subset of those founds in the \code{data} field of \code{cohort_data} (i.e.,
all IDs in a covariate dataset must also be present in the cohort dataset.)
A covariate measurement stored in the \code{data} field of an element
\code{timeDepCovData}
of \code{cov_data} will be ignored
if its date is either strictly before the subject's index date stored in the
\code{data} field of \code{cohort_data} or strictly after the subject's end of
follow-up date stored in the \code{data} field of \code{cohort_data}. (i.e., only
covariate measurements collected between the start and end of follow-up will be
retained).
}
}
}

\keyword{data}
\section{Super class}{
\code{\link[LtAtStructuR:LtAtData]{LtAtStructuR::LtAtData}} -> \code{LtAtDataInst}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LtAtDataInst$new()}}
\item \href{#method-copyLtAtData}{\code{LtAtDataInst$copyLtAtData()}}
\item \href{#method-construct}{\code{LtAtDataInst$construct()}}
\item \href{#method-setAlgoOptions}{\code{LtAtDataInst$setAlgoOptions()}}
\item \href{#method-assignAC}{\code{LtAtDataInst$assignAC()}}
\item \href{#method-assignL}{\code{LtAtDataInst$assignL()}}
\item \href{#method-cleanUp}{\code{LtAtDataInst$cleanUp()}}
\item \href{#method-clone}{\code{LtAtDataInst$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="LtAtStructuR" data-topic="LtAtData" data-id="addSpec">}\href{../../LtAtStructuR/html/LtAtData.html#method-addSpec}{\code{LtAtStructuR::LtAtData$addSpec()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="LtAtStructuR" data-topic="LtAtData" data-id="createIntervals">}\href{../../LtAtStructuR/html/LtAtData.html#method-createIntervals}{\code{LtAtStructuR::LtAtData$createIntervals()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="LtAtStructuR" data-topic="LtAtData" data-id="imputeL">}\href{../../LtAtStructuR/html/LtAtData.html#method-imputeL}{\code{LtAtStructuR::LtAtData$imputeL()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LtAtDataInst$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copyLtAtData"></a>}}
\if{latex}{\out{\hypertarget{method-copyLtAtData}{}}}
\subsection{Method \code{copyLtAtData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LtAtDataInst$copyLtAtData(LtAtObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-construct"></a>}}
\if{latex}{\out{\hypertarget{method-construct}{}}}
\subsection{Method \code{construct()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LtAtDataInst$construct(time_unit, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setAlgoOptions"></a>}}
\if{latex}{\out{\hypertarget{method-setAlgoOptions}{}}}
\subsection{Method \code{setAlgoOptions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LtAtDataInst$setAlgoOptions(time_unit, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assignAC"></a>}}
\if{latex}{\out{\hypertarget{method-assignAC}{}}}
\subsection{Method \code{assignAC()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LtAtDataInst$assignAC()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assignL"></a>}}
\if{latex}{\out{\hypertarget{method-assignL}{}}}
\subsection{Method \code{assignL()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LtAtDataInst$assignL()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleanUp"></a>}}
\if{latex}{\out{\hypertarget{method-cleanUp}{}}}
\subsection{Method \code{cleanUp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LtAtDataInst$cleanUp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LtAtDataInst$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
