% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataConstruction_workflow.R
\name{setCohort}
\alias{setCohort}
\title{Definition of cohort dataset}
\usage{
setCohort(
  data,
  IDvar,
  index_date,
  EOF_date,
  EOF_type,
  Y_name,
  L0,
  L0_timeIndep = NA
)
}
\arguments{
\item{data}{\code{data.table} containing the input cohort dataset to be
wrapped in and processed. The table must contain a single row for
each subject in the cohort. Cannot have columns named 'IDvar',
'index_date', 'EOF_date', 'EOF_type', or 'L0'.}

\item{IDvar}{\code{character} providing the name of the column of
\code{data} that contains the unique subject identifier.}

\item{index_date}{\code{character} providing the name of the column of
\code{data} that contains the date of study entry.}

\item{EOF_date}{\code{character} providing the name of the column of
\code{data} that contains the date of end of follow-up. All observations
with the end of follow-up date equal to the study entry date will be ignored
(i.e., excluded from the cohort).}

\item{EOF_type}{\code{character} providing the name of the column of
\code{data} corresponding to the reason for end of follow-up.}

\item{Y_name}{\code{character} or \code{integer}  providing the unique value
in column \code{EOF_type} that encodes the end of follow-up due to failure
(i.e., occurrence of the outcome event of interest).}

\item{L0}{vector of \code{character} providing the names of the columns of
\code{data} that contain baseline covariate measurements.}

\item{L0_timeIndep}{named list specifying, for each time-independent covariates in
\code{L0}, a sublist with only the following three named elements:
\enumerate{
\item \code{categorical}: specifies whether the covariate is
continuous ('FALSE') or categorical ('TRUE'). Cannot be missing.
\item \code{impute}: specifies the imputation method for missing
measurements: 'default', 'mean', 'mode', 'median'. If missing, imputation with
the 'mean' and 'mode' is used for continuous and categorical covariates,
respectively. Imputation with 'mean', 'mode', or 'median' is based on
measurements from subjects with observed covariate values in \code{data}.
'mean' and 'median'
can only be used for continuous covariates. 'mode' can only be used for
categorical covariates. Imputation with 'default' replaces missing values with
0 if the covariate is numeric and with 'Unknown' otherwise.
\item \code{impute_default_level} imputation value to be used when
the imputation method is 'default'. The value must be
a length 1 \code{character} (resp. \code{numeric}) for a covariate encoded by a
\code{character} (resp. \code{numeric}) vector.
If missing, the default values 0 and 'Unknown' are used for continuous and
categorical covariates, respectively.
}
Each element of the list \code{L0_timeIndep} must be named with the time-independent
covariate in \code{L0} to which the sublist information applies.}
}
\value{
\code{cohortData} object
}
\description{
The cohort dataset specifies for each subject in the cohort:
\enumerate{
\item a unique subject identifier,
\item the date of study entry,
\item the date of end of follow-up,
\item the reason for end of follow-up (failure or right-censoring), and
\item baseline measurements of time-dependent or time-independent covariates.
}
}
\examples{
cohort <- setCohort(cohortDT, "ID", "IndexDate", "EOFDate", "EOFtype",
                    "AMI", c("ageEntry", "sex", "race", "A1c", "eGFR"),
                    list("ageEntry"=list("categorical"=FALSE,
                                         "impute"=NA,
                                         "impute_default_level"=NA),
                         "sex"=list("categorical"=TRUE,
                                    "impute"=NA,
                                    "impute_default_level"=NA),
                         "race"=list("categorical"=TRUE,
                                     "impute"=NA,
                                     "impute_default_level"=NA)) ) 
}
\seealso{
\link{cohortData}
}
\keyword{data}
