% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataConstruction_workflow.R
\name{setExposure}
\alias{setExposure}
\title{Definition of exposure dataset}
\usage{
setExposure(data, IDvar, start_date, end_date, exp_level = NA, exp_ref = NA)
}
\arguments{
\item{data}{\code{data.table} containing the input exposure dataset to be
wrapped in and processed. The table can contain multiple rows per
subject. There should be no row for subjects who are only
exposed to the reference exposure level during follow-up.
Exposure levels must be encoded by a character vector or an
integer vector. There cannot be any overlapping exposure
episodes. Cannot contain missing values. Cannot have columns
named 'IDvar', 'start_date', 'end_date', 'exposure', or 'exp_level'.}

\item{IDvar}{\code{character} providing the name of the column of
\code{data} that contains the unique subject identifier.}

\item{start_date}{\code{character} providing the name of the column of
\code{data} that contains the exposure episode start date.}

\item{end_date}{\code{character} providing the name of the column of
\code{data} that contains the exposure episode end date.}

\item{exp_level}{\code{character} providing the name of the column of
\code{data} that contains the non-reference exposure level.
Can be missing only if \code{exp_ref} is also missing. If missing,
the exposure is assumed to be binary and its reference level is
encoded by 0.}

\item{exp_ref}{\code{character} or \code{integer} identifying the exposure
reference level. Cannot be a value of the exp_level column of
data. Can be missing only if \code{exp_level} is also missing.}
}
\value{
\code{expData} object
}
\description{
The exposure dataset specifies all follow-up time intervals during which a
study subject is exposed to an exposure level other than a reference level
specified by the analyst. For each episode of exposure to a non-reference
level, the table specifies:
\enumerate{
\item a unique subject identifier,
\item the exposure episode start date,
\item the exposure episode end date, and
\item the non-reference exposure level (required only when the exposure is not
binary).
}
}
\examples{
exposure <- setExposure(expDT, "ID", "startA", "endA")
}
\seealso{
\link{expData}
}
\keyword{data}
