% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataConstruction_workflow.R
\name{setInstantExposure}
\alias{setInstantExposure}
\title{Definition of instant exposure dataset}
\usage{
setInstantExposure(data, IDvar, exp_date, exp_level = NA)
}
\arguments{
\item{data}{\code{data.table} containing the input exposure dataset to be
wrapped in and processed. The table can contain multiple rows per
subject. There should be no row for subjects who are only
exposed to the reference exposure level during follow-up.
Exposure levels must be encoded by one or more character,
integer, or numeric vector(s). Multiple rows per subject and date
are permitted. Cannot contain missing values. Cannot have columns
named 'IDvar', 'start_date', or 'exp_level'.}

\item{IDvar}{\code{character} providing the name of the column of
\code{data} that contains the unique subject identifier.}

\item{exp_date}{\code{character} providing the name of the column of
\code{data} that contains the exposure date.}

\item{exp_level}{\code{character} vector providing the name(s) of the column(s) of
\code{data} that encode(s) the non-reference exposure level(s).
Can be missing if there is only one non-reference exposure level. If missing,
the exposure is assumed to be binary and its reference level is
encoded by 0.}
}
\value{
\code{instExpData} object
}
\description{
The instant exposure dataset specifies all follow-up dates when a
study subject is exposed to a non-reference exposure level.
For each instanteneous exposures, the table specifies:
\enumerate{
\item a unique subject identifier,
\item the exposure date,
\item the exposure level (required only when the exposure is not
binary).
}
}
\examples{
exposure <- setInstantExposure(expDT6, "ID", "fillDate", c("D.t","Q.t"))
}
\seealso{
\link{instantExpData}
}
\keyword{data}
