% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataClass.R
\docType{class}
\name{timeDepCovData}
\alias{timeDepCovData}
\title{Data Storage Class for a time-dependent covariate dataset}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{timeDepCovData} object
}
\description{
Class that defines the standard format for a dataset that encodes all follow-up measurements for
a single time-dependent covariate, i.e. the table that specifies for each subject in the
cohort: 1) a unique subject identifier,  2) the date of the covariate measurement, 3) the
covariate value.
}
\section{Fields}{

\describe{
\item{\code{data}:}{\code{data.table} containing
an input covariate dataset to be wrapped in and processed.
The table can contain multiple rows per subject. There
should be no row for subjects with no follow-up covariate measurements. The table
must contain at most one covariate measurement on a given date for any given subject.
Cannot contain missing values. Covariate values must be encoded by a character or
numeric vector (e.g., factors are not allowed).
Cannot have columns named 'IDvar', 'L_date', or 'L_name'.
}
\item{\code{type}:}{\code{character} specifying the covariate type: 'binary monotone
increasing' (e.g., history of a diagnosis or procedure), 'interval' (e.g., hospital
stay or prescription coverage), 'sporadic' (e.g., laboratory measurements),
'indicator' (e.g., occurrence of a repeatable event).
}
\item{\code{IDvar}:}{\code{character} providing the name of the column of
\code{data} that contains the unique subject identifier.
}
\item{\code{L_date}:}{\code{character} providing the name of the column of
\code{data} that contains the date of the follow-up covariate measurement.
}
\item{\code{L_name}:}{\code{character} providing the name of
the column of \code{data} that contains the covariate values. For all covariate types,
\code{L_name} must also be the name of the column of the cohort dataset that contains
the baseline measurements of the time-dependent covariate. Baseline measurements of a
covariate of type 'binary monotone increasing' can only be encoded with values 0 and 1 in
the cohort dataset. All values in the column \code{L_name} of \code{data} must be set
to 1 for a covariate of type 'binary monotone increasing'.
The column \code{L_name} in \code{data} cannot contain
the value 'None' for a covariate of type 'indicator' that is \code{character}.
The column \code{L_name} in \code{data} and in the cohort dataset cannot contain
the value 0 for a covariate of type 'indicator' that is \code{numeric}.
}
\item{\code{categorical}:}{\code{logical} indicating whether the covariate is continuous
('FALSE') or categorical ('TRUE'). Must be 'TRUE' for a covariate of type 'binary
monotone increasing' or 'indicator'. Cannot be missing.
}
\item{\code{impute}:}{\code{character} specifying imputation method for missing baseline
measurements: 'default', 'mean', 'mode', 'median'. If missing, imputation with
the 'mean' and 'mode' is used for continuous and categorical covariates,
respectively. Imputation with 'mean', 'mode', or 'median' is based on baseline
measurements from subjects with observed baseline covariate values (stored in the
cohort dataset). 'mean' and 'median' can only be used for continuous covariates.
'mode' can only be used for categorical covariates. Imputation with 'default'
replaces missing values with 0 if the covariate is numeric and with 'Unknown'
otherwise. Ignored for a covariate of type 'binary monotone increasing', 'interval',
or 'indicator'.
}
\item{\code{impute_default_level}:}{ \code{character} or \code{numeric} specifying the
imputation value to be used when \code{impute}='default'. The value must be
a length 1 \code{character} (resp. \code{numeric}) for a covariate encoded by a
\code{character} (resp. \code{numeric}) vector.
If missing, the default
values 0 and 'Unknown' are used for numeric and character covariates,
respectively. Ignored for a
covariate of type 'binary monotone increasing', 'interval', or 'indicator'.
}
\item{\code{acute_change}:}{\code{logical} indicating whether a covariate measurement
collected on the date of an exposure change can be impacted by the change.
The default value 'FALSE' indicates that the covariate measurement can be assumed to
have preceded and possibly triggered the change in exposure. Cannot be missing.
}
}
}

\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{timeDepCovData$new()}}
\item \href{#method-checkAgainst}{\code{timeDepCovData$checkAgainst()}}
\item \href{#method-clone}{\code{timeDepCovData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{timeDepCovData$new(
  data,
  type,
  IDvar,
  L_date,
  L_name,
  categorical,
  impute = NA,
  impute_default_level = NA,
  acute_change = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkAgainst"></a>}}
\if{latex}{\out{\hypertarget{method-checkAgainst}{}}}
\subsection{Method \code{checkAgainst()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{timeDepCovData$checkAgainst(otherData)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{timeDepCovData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
