% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataConstruction_workflow.R
\name{construct}
\alias{construct}
\title{Main function of the package}
\usage{
construct(
  LtAtspec,
  time_unit,
  first_exp_rule = 1,
  exp_threshold = 0.5,
  format = "standard",
  dates = FALSE
)
}
\arguments{
\item{LtAtspec}{specifies an object of class \code{LtAtData} created by
assembling cohort, exposure, and time-dependent covariate
data objects. Cannot be missing.}

\item{time_unit}{specifies the unit of time (expressed in days) for
discretizing follow-time into consecutive time intervals
of the same length. Must specify a strictly positive
integer value. Cannot be missing.}

\item{first_exp_rule}{specifies a binary value that indicates which of
two algorithms is used to assign
the first change in exposure status during follow-up.
Default value is 1.}

\item{exp_threshold}{specifies a threshold used to assign an exposure level
at each time interval. Must be a value
between 0 and 1. Default value is 0.5.}

\item{format}{specifies the format of the output data set. Current possible
values are "standard", or "MSM SAS macro". Default value is
"standard".}

\item{dates}{indicates whether the output dataset should contain dates.
Default value is FALSE.}
}
\value{
\code{data.table} object
}
\description{
Implements a coarsening algorithm to map input exposure and covariate data
into a structured analytic data set that encodes complex, discrete-time,
longitudinal data. \cr
The encoded data on each experimental unit is a temporally ordered sequence
of covariates L(t) and exposures A(t) where covariates at time t occur
before the exposure A(t) a time t:\cr
O=(L(0),A(0),...,L(K),A(K),L(K+1)),  \cr
where each t represents one of the consecutive time intervals that result
from discretizing, separately for each experimental unit, the elapsed time
between an index date and end of follow-up date using a fixed unit of time
specified by the analyst. \cr
In particular, the resulting data set can be used for the evaluation of
the causal effects of various exposure regimens on a time-to-event
(survival) outcome subject to right-censoring in a cohort study.
}
\examples{
## Define one cohort dataset, one exposure dataset, and one or more covariate
## datasets
cohort <- setCohort(cohortDT, "ID", "IndexDate", "EOFDate", "EOFtype",
                    "AMI", c("ageEntry", "sex", "race", "A1c", "eGFR"),
                    list("ageEntry"=list("categorical"=FALSE,
                                         "impute"=NA,
                                         "impute_default_level"=NA),
                         "sex"=list("categorical"=TRUE,
                                    "impute"=NA,
                                    "impute_default_level"=NA),
                         "race"=list("categorical"=TRUE,
                                     "impute"=NA,
                                     "impute_default_level"=NA)) )
exposure <- setExposure(expDT, "ID", "startA", "endA")
covariate1 <- setCovariate(a1cDT, "sporadic", "ID", "A1cDate", "A1c",
                           categorical = FALSE)
covariate2 <- setCovariate(egfrDT, "sporadic", "ID", "eGFRDate", "eGFR",
                           categorical = TRUE)

## Gather each input dataset into a single object that specifies the content of
## the output dataset to be constructed
LtAt.specification <- cohort + exposure + covariate1 + covariate2

## Construct the output dataset
LtAt.data <- construct(LtAt.specification, time_unit = 15, first_exp_rule = 1,
                       exp_threshold = 0.75)

}
\seealso{
\link{LtAtData}
}
\keyword{function}
